requirements.txt
streamlit
pycryptodome
Pillow
mysql-connector-python
sqlalchemy